# packages/contracts/generated/spring/src/main/java/tv/lumo/api/generated/api/UserdataApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package tv.lumo.api.generated.api;

import tv.lumo.api.generated.model.AddFavoriteRequest;
import tv.lumo.api.generated.model.CreateFavoriteGroupRequest;
import tv.lumo.api.generated.model.Favorite;
import tv.lumo.api.generated.model.FavoriteGroup;
import tv.lumo.api.generated.model.FavoriteGroupList;
import tv.lumo.api.generated.model.FavoriteList;
import tv.lumo.api.generated.model.PlaybackProgress;
import tv.lumo.api.generated.model.Problem;
import tv.lumo.api.generated.model.SaveProgressRequest;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
@Validated
public interface UserdataApi {

    /**
     * POST /me/favorites : Add a channel to a favourite group
     * &#x60;group_id&#x60; may be omitted: the channel then lands in the account&#39;s default group, which is created on the first add. 
     *
     * @param addFavoriteRequest  (required)
     * @return The created favourite. (status code 201)
     *         or The request is malformed or fails validation (&#x60;VALIDATION_FAILED&#x60;). (status code 400)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or The channel or the group does not belong to the caller (&#x60;CHANNEL_NOT_FOUND&#x60;, &#x60;FAVORITE_GROUP_NOT_FOUND&#x60;).  (status code 404)
     *         or This channel is already in that group (&#x60;FAVORITE_ALREADY_EXISTS&#x60;). (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/me/favorites",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Favorite> addFavorite(
         @Valid @RequestBody AddFavoriteRequest addFavoriteRequest
    );


    /**
     * POST /me/favorite-groups : Create a favourite group
     *
     * @param createFavoriteGroupRequest  (required)
     * @return The created group. (status code 201)
     *         or The request is malformed or fails validation (&#x60;VALIDATION_FAILED&#x60;). (status code 400)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or A group with this name already exists (&#x60;FAVORITE_GROUP_ALREADY_EXISTS&#x60;). (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/me/favorite-groups",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<FavoriteGroup> createFavoriteGroup(
         @Valid @RequestBody CreateFavoriteGroupRequest createFavoriteGroupRequest
    );


    /**
     * GET /me/favorite-groups : The user&#39;s favourite groups
     *
     * @return Groups ordered by &#x60;position&#x60;. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/me/favorite-groups",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<FavoriteGroupList> listFavoriteGroups(
        
    );


    /**
     * GET /me/favorites : The user&#39;s favourites
     *
     * @param groupId Restrict to one favourite group. (optional)
     * @return Favourites ordered by group then &#x60;position&#x60;. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/me/favorites",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<FavoriteList> listFavorites(
         @Valid @RequestParam(value = "groupId", required = false) @Nullable UUID groupId
    );


    /**
     * DELETE /me/favorites/{id} : Remove a favourite
     *
     * @param id Resource identifier. (required)
     * @return Favourite removed. (status code 204)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or No such favourite on this account (&#x60;FAVORITE_NOT_FOUND&#x60;). (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/me/favorites/{id}",
        produces = { "application/problem+json" }
    )
    
    ResponseEntity<Void> removeFavorite(
         @PathVariable("id") UUID id
    );


    /**
     * PUT /me/progress : Save playback progress for a VOD item or an episode
     * Idempotent upsert keyed on &#x60;(item_type, item_ref)&#x60; for the caller.  Live channels have no progress. Sending &#x60;item_type&#x60; for a live channel is a client bug, not a supported case. 
     *
     * @param saveProgressRequest  (required)
     * @return The stored progress. (status code 200)
     *         or The request is malformed or fails validation (&#x60;VALIDATION_FAILED&#x60;). (status code 400)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/me/progress",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<PlaybackProgress> saveProgress(
         @Valid @RequestBody SaveProgressRequest saveProgressRequest
    );

}