# packages/contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Lumo TV API
  version: 1.0.0
  description: |
    Backend for the Lumo TV IPTV player (`lumo-api`).

    Lumo hosts, provides and resells **no content**. The user brings their own
    source — an M3U playlist or an Xtream Codes account. This API manipulates
    metadata only: catalogue, guide, favourites, playback progress, access
    rights. **Media never transits through this infrastructure**: the player
    opens the stream directly against the user's own IPTV server.

    ## Authentication

    All operations require `Authorization: Bearer <access_token>` unless the
    operation declares `security: []`. Access tokens are short-lived JWTs
    (15 minutes). Refresh tokens are opaque, rotated on every use, and bound to
    a `device`. Presenting an already-consumed refresh token revokes the whole
    token chain for that device — that is the theft-detection mechanism.

    ## Errors

    Every error response is RFC 7807 `application/problem+json` and carries a
    stable, machine-readable `code`. Clients switch on `code` and render their
    own localised message (FR/EN). `detail` is meant for logs and is never
    displayed raw to a user.

    ## Asynchronous ingestion

    Parsing is always server-side. `POST /sources` performs a fast synchronous
    reachability and credential check, then returns `202` with the source in
    `PENDING`; the full catalogue ingestion runs in the background. Clients poll
...
    blocks an HTTP request.
  contact:
...
servers:
...
security:
  - bearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /me/favorites:
    get:
      tags: [userdata]
      operationId: listFavorites
      summary: The user's favourites
      parameters:
        - name: groupId
          in: query
          required: false
          description: Restrict to one favourite group.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Favourites ordered by group then `position`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteList'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [userdata]
      operationId: addFavorite
      summary: Add a channel to a favourite group
      description: |
        `group_id` may be omitted: the channel then lands in the account's
        default group, which is created on the first add.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavoriteRequest'
      responses:
        '201':
          description: The created favourite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: |
            The channel or the group does not belong to the caller
            (`CHANNEL_NOT_FOUND`, `FAVORITE_GROUP_NOT_FOUND`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: This channel is already in that group (`FAVORITE_ALREADY_EXISTS`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /me/favorites/{id}:
    delete:
      tags: [userdata]
      operationId: removeFavorite
      summary: Remove a favourite
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Favourite removed.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: No such favourite on this account (`FAVORITE_NOT_FOUND`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /me/favorite-groups:
    get:
      tags: [userdata]
      operationId: listFavoriteGroups
      summary: The user's favourite groups
      responses:
        '200':
          description: Groups ordered by `position`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteGroupList'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [userdata]
      operationId: createFavoriteGroup
      summary: Create a favourite group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFavoriteGroupRequest'
      responses:
        '201':
          description: The created group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteGroup'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: A group with this name already exists (`FAVORITE_GROUP_ALREADY_EXISTS`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /me/progress:
    put:
      tags: [userdata]
      operationId: saveProgress
      summary: Save playback progress for a VOD item or an episode
      description: |
        Idempotent upsert keyed on `(item_type, item_ref)` for the caller.

        Live channels have no progress. Sending `item_type` for a live channel is
        a client bug, not a supported case.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProgressRequest'
      responses:
        '200':
          description: The stored progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackProgress'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

# ==============================================================================
# COMPONENTS
# ==============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access token (15 minutes) obtained from `/auth/login`,
        `/auth/register`, `/auth/oauth/google`, `/auth/refresh` or
        `/auth/device/token`.

  # ---------------------------------------------------------- parameters ------
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource identifier.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation (`VALIDATION_FAILED`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: |
        Missing, malformed or expired access token (`UNAUTHENTICATED`,
        `ACCESS_TOKEN_EXPIRED`). On `ACCESS_TOKEN_EXPIRED` the client refreshes
        once and replays the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem detail. Every error response in this API uses this
        shape with `Content-Type: application/problem+json`.

        Clients switch on `code`, never on `title` or `detail`. `detail` is
        written for logs and support, is not localised, and is never displayed
        raw to a user.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem type.
          example: https://lumo.tv/errors/source-auth-failed
        title:
          type: string
          description: Short, human-readable, English, non-localised summary.
          example: Source authentication failed
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: HTTP status code, repeated for convenience.
          example: 422
        code:
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
          description: |
            Stable machine-readable code. **This is the only field a client is
            allowed to branch on.** Clients must handle an unrecognised value
            gracefully rather than crash: new codes may be added within v1.
        detail:
          type: string
          nullable: true
          description: |
            Diagnostic explanation for logs and support. Never rendered raw in
            the UI. Never contains a stream URL, an Xtream password or a token.
          example: Connection timed out after 10s
        instance:
          type: string
          format: uri-reference
          nullable: true
          description: Path of the request that produced this problem.
          example: /v1/sources
        errors:
          type: array
          nullable: true
          description: Per-field details, present on validation failures only.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      description: One field-level validation failure.
      required: [field, code]
      properties:
        field:
          type: string
          description: JSON pointer to the offending property, relative to the request body.
          example: /password
        code:
          type: string
          description: Stable machine-readable reason.
          example: TOO_SHORT
        detail:
          type: string
          nullable: true
          description: Diagnostic explanation. Not localised, not for display.
...
    ErrorCode:
      type: string
      description: |
        Every stable error code this API emits, ingestion codes included.

        Grouped by concern:

        - **Generic** — `VALIDATION_FAILED`, `UNAUTHENTICATED`,
          `ACCESS_TOKEN_EXPIRED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`,
          `RATE_LIMITED`, `INTERNAL_ERROR`.
        - **Account and sign-in** — `EMAIL_ALREADY_REGISTERED`,
          `INVALID_CREDENTIALS`, `PASSWORD_TOO_WEAK`, `OAUTH_TOKEN_INVALID`,
          `VERIFICATION_TOKEN_INVALID`, `VERIFICATION_TOKEN_EXPIRED`,
          `RESET_TOKEN_INVALID`, `RESET_TOKEN_EXPIRED`.
        - **Sessions** — `REFRESH_TOKEN_INVALID`, `REFRESH_TOKEN_REUSED`,
          `DEVICE_NOT_FOUND`.
        - **TV activation (RFC 8628)** — `AUTHORIZATION_PENDING`, `SLOW_DOWN`,
          `ACCESS_DENIED`, `EXPIRED_TOKEN`, `DEVICE_CODE_NOT_FOUND`,
          `DEVICE_CODE_EXPIRED`, `DEVICE_CODE_ALREADY_USED`.
        - **Sources and ingestion** — `SOURCE_NOT_FOUND`, `SOURCE_NOT_READY`,
          `SOURCE_SYNC_IN_PROGRESS`, `SOURCE_SYNC_RATE_LIMITED`, plus every
          `IngestionErrorCode`.
        - **Catalogue and user data** — `CHANNEL_NOT_FOUND`,
          `FAVORITE_NOT_FOUND`, `FAVORITE_ALREADY_EXISTS`,
          `FAVORITE_GROUP_NOT_FOUND`, `FAVORITE_GROUP_ALREADY_EXISTS`.
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - ACCESS_TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
        - EMAIL_ALREADY_REGISTERED
        - INVALID_CREDENTIALS
        - PASSWORD_TOO_WEAK
        - OAUTH_TOKEN_INVALID
        - VERIFICATION_TOKEN_INVALID
        - VERIFICATION_TOKEN_EXPIRED
        - RESET_TOKEN_INVALID
        - RESET_TOKEN_EXPIRED
        - REFRESH_TOKEN_INVALID
        - REFRESH_TOKEN_REUSED
        - DEVICE_NOT_FOUND
        - AUTHORIZATION_PENDING
        - SLOW_DOWN
        - ACCESS_DENIED
        - EXPIRED_TOKEN
        - DEVICE_CODE_NOT_FOUND
        - DEVICE_CODE_EXPIRED
        - DEVICE_CODE_ALREADY_USED
        - SOURCE_NOT_FOUND
        - SOURCE_NOT_READY
        - SOURCE_SYNC_IN_PROGRESS
        - SOURCE_SYNC_RATE_LIMITED
        - SOURCE_UNREACHABLE
        - SOURCE_AUTH_FAILED
        - SOURCE_EXPIRED
        - SOURCE_MAX_CONNECTIONS
        - SOURCE_INVALID_FORMAT
        - SOURCE_EMPTY
        - SOURCE_TOO_LARGE
        - CHANNEL_NOT_FOUND
        - FAVORITE_NOT_FOUND
        - FAVORITE_ALREADY_EXISTS
        - FAVORITE_GROUP_NOT_FOUND
        - FAVORITE_GROUP_ALREADY_EXISTS

    # ---- enums -------------------------------------------------------------
...
    ProgressItemType:
      type: string
      description: |
        What a `playback_progress` row points at. Live channels have no
        progress, so there is no `LIVE` value here.
      enum: [VOD, EPISODE]

    # ---- auth --------------------------------------------------------------
...
    FavoriteGroup:
      type: object
      description: |
        A user-defined grouping of favourites. A default group is created on the
        first add.
      required: [id, name, position]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        position:
          type: integer
          format: int32
    FavoriteGroupList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteGroup'
    CreateFavoriteGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        position:
          type: integer
          format: int32
          nullable: true
          description: Appended last when omitted.
    Favorite:
      type: object
      description: |
        A channel bookmarked by the user. Survives a re-synchronisation as long
        as the channel keeps its identity within the source.
      required: [id, group_id, source_id, channel_id, position]
      properties:
        id:
          type: string
          format: uuid
        group_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        channel_id:
          type: string
          format: uuid
        position:
          type: integer
          format: int32
    FavoriteList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Favorite'
    AddFavoriteRequest:
      type: object
      description: |
        `source_id` is not accepted: it is derived from the channel, so the two
        cannot disagree.
      required: [channel_id]
      properties:
        channel_id:
          type: string
          format: uuid
        group_id:
          type: string
          format: uuid
          nullable: true
          description: Defaults to the account's default group, created on first use.
        position:
          type: integer
          format: int32
          nullable: true
          description: Appended last when omitted.
    PlaybackProgress:
      type: object
      description: Where the user stopped watching a VOD item or an episode.
      required: [id, item_type, item_ref, position_ms, updated_at]
      properties:
        id:
          type: string
          format: uuid
        item_type:
          $ref: '#/components/schemas/ProgressItemType'
        item_ref:
          type: string
          description: Identifier of the item within its source.
        position_ms:
          type: integer
          format: int64
          minimum: 0
          description: Playback position, in milliseconds.
        duration_ms:
          type: integer
          format: int64
          nullable: true
          minimum: 0
          description: Total duration, when known.
        updated_at:
          type: string
          format: date-time
    SaveProgressRequest:
      type: object
      description: Upsert keyed on `(item_type, item_ref)` for the caller.
      required: [item_type, item_ref, position_ms]
      properties:
        item_type:
          $ref: '#/components/schemas/ProgressItemType'
        item_ref:
          type: string
          minLength: 1
          maxLength: 200
        position_ms:
          type: integer
          format: int64
          minimum: 0
        duration_ms:
          type: integer
          format: int64
          nullable: true
          minimum: 0